requirements/base.txt
# Base requirements for all environments
Django>=5.0.0,<5.1.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.5
python-decouple>=3.8
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
drf-yasg>=1.21.7
django-filter>=23.3
Pillow>=10.0.0
gunicorn>=21.2.0
whitenoise>=6.6.0
celery>=5.3.0
redis>=5.0.0
